+++ changelog.md
@@ -1,5 +1,6 @@
 <h2><u><strong>Changelogs</strong></u></h2>
 
-1. added rickroll in `console`
-1. using `dynamic` editor component
-1. made onSave and onExit props on editor
+1. create a memoized component for MonacoEditor.
+1. create a memoized component for preview.
+1. fixed the MonacoEditor overlaps the preview bug.
+1. added onExit and onSave on click for mobile devices.